[refac] refactor service package structure to unify port definitions
