+++ edukit-core/src/main/java/com/edukit/core/common/port/AIService.java
@@ -1,4 +1,4 @@
-package com.edukit.core.common.service.port;
+package com.edukit.core.common.port;
 
 import com.edukit.core.common.service.response.OpenAIVersionResponse;
 import reactor.core.publisher.Flux;
