+++ edukit-core/src/main/java/com/edukit/core/common/port/EmailService.java
@@ -1,4 +1,4 @@
-package com.edukit.core.common.service.port;
+package com.edukit.core.common.port;
 
 public interface EmailService {
