+++ edukit-core/src/main/java/com/edukit/core/common/port/RedisService.java
@@ -1,4 +1,4 @@
-package com.edukit.core.common.service.port;
+package com.edukit.core.common.port;
 
 import java.time.Duration;
